fixup! Merge remote-tracking branch 'origin/master' into enable_sparse

+++ ADOL-C/include/adolc/valuetape/sparseinfos.h
@@ -38,10 +38,10 @@ struct ADOLC_API SparseJacInfos {
 
   void setJP(std::vector<uint *> &&JPIn) {
     // delete JP
-    for (auto& j: JP_)
+    for (auto &j : JP_)
       delete[] j;
     JP_ = std::move(JPIn);
-    depen_ = JP_.size();
+    depen_ = static_cast<int>(JP_.size());
   }
   std::vector<uint *> &getJP() { return JP_; }
   void initColoring(int dimOut, int dimIn);
@@ -88,7 +88,7 @@ public:
   std::vector<uint *> &getHP() { return HP_; }
   void setHP(int indep, std::vector<uint *> &&HPIn) {
     // delete HP_
-    for (auto& h: HP_)
+    for (auto &h : HP_)
       delete[] h;
     indep_ = indep;
     HP_ = std::move(HPIn);
